Remove redirects
Redirects were for pages that were moved on Bitcoin.org.  New site has
no legacy links and so doesn't need redirects.

Now that we have no redirects, remove custom redircets plugin so that
when we do eventually need redirects for this page, we use an upstream
maintained plugin like redirect_from or redirect_to as used on
BitcoinCore.org.